refactor(core): migrate domain types to shared layer
