+++ packages/core/src/shared/types/document.ts
@@ -1,0 +1,12 @@
+export type DocumentStatus = 'pending' | 'processing' | 'completed' | 'failed';
+
+export interface DocumentEntity {
+  id: string;
+  userId: string;
+  name: string;
+  storagePath: string;
+  fileSize: number;
+  status: DocumentStatus;
+  createdAt: number;
+  updatedAt: number;
+}
