+++ packages/core/src/types/document.ts
@@ -1,10 +1,1 @@
-export interface DocumentEntity {
-	id: string;
-	userId: string;
-	name: string;
-	storagePath: string;
-	fileSize: number;
-	status: "pending" | "processing" | "completed" | "failed";
-	createdAt: number;
-	updatedAt: number;
-}
+export * from '../shared/types/document.js';
